Grammar

+++ fuzzing_campaigns/run_clustering.sh
@@ -26,30 +26,70 @@ for ndrivers in "${NUM_OF_DRIVERS[@]}"; do
     for napis in "${NUM_OF_APIs[@]}"; do        
         for project in "${PROJECTS[@]}"; do
 
-            PROJECT_WORKDIR="./workdir_${ndrivers}_${napis}/${project}"
-            DRIVER_FOLDER="${PROJECT_WORKDIR}/drivers"
-	    if [ -d "$DRIVER_FOLDER" ]; then
-            	DRIVER_NAMES="$(find ${DRIVER_FOLDER} -type f -executable)"
-	    else 
-		continue
-	    fi
-            CRASHES=${PROJECT_WORKDIR}/crashes
-            
-            for i in $( eval echo {1..$ITERATIONS} ); do
-                for driver_name in $DRIVER_NAMES; do
-                    driver_name=$(basename $driver_name)
-                    rm -Rf ${CRASHES}/${driver_name} || true
-                    mkdir -p ${CRASHES}/${driver_name}
+            if [[ -z ${GRAMMAR_MODE} ]]; then
 
-                    DRIVER_CRASHES="${PROJECT_WORKDIR}/results/iter_${i}/crashes/${driver_name}"
-                    [ "$(ls -A ${DRIVER_CRASHES})" ] \
-                        && cp ${DRIVER_CRASHES}/* ${CRASHES}/${driver_name} || echo "No crashes for ${project}/${driver_name} on iter ${i}"
+                PROJECT_WORKDIR="./workdir_${ndrivers}_${napis}/${project}"
+                DRIVER_FOLDER="${PROJECT_WORKDIR}/drivers"
+                if [ -d "$DRIVER_FOLDER" ]; then
+                    DRIVER_NAMES="$(find ${DRIVER_FOLDER} -type f -executable)"
+                else 
+                    continue
+                fi
+                CRASHES=${PROJECT_WORKDIR}/crashes
+                
+                for i in $( eval echo {1..$ITERATIONS} ); do
+                    for driver_name in $DRIVER_NAMES; do
+                        driver_name=$(basename $driver_name)
+                        rm -Rf ${CRASHES}/${driver_name} || true
+                        mkdir -p ${CRASHES}/${driver_name}
+
+                        DRIVER_CRASHES="${PROJECT_WORKDIR}/results/iter_${i}/crashes/${driver_name}"
+                        [ "$(ls -A ${DRIVER_CRASHES})" ] \
+                            && cp ${DRIVER_CRASHES}/* ${CRASHES}/${driver_name} || echo "No crashes for ${project}/${driver_name} on iter ${i}"
+                    done
                 done
-            done
-            
-            PROJECT_FOLDER="/workspaces/libfuzz/fuzzing_campaigns/workdir_${ndrivers}_${napis}/${project}"
-            docker run --privileged --env TARGET=${project} --env TARGET_WORKDIR=${PROJECT_FOLDER} \
-                    -v $(pwd)/..:/workspaces/libfuzz "${IMG_NAME}-${project}" || true
+
+                PROJECT_FOLDER="/workspaces/libfuzz/fuzzing_campaigns/workdir_${ndrivers}_${napis}/${project}"
+                docker run --privileged --env TARGET=${project} \
+                        --env TARGET_WORKDIR=${PROJECT_FOLDER} \
+                        --env GRAMMAR_MODE=${GRAMMAR_MODE} \
+                        -v $(pwd)/..:/workspaces/libfuzz "${IMG_NAME}-${project}" || true
+
+            else
+
+                for i in $( eval echo {1..$ITERATIONS} ); do
+
+                    PROJECT_WORKDIR="./workdir_${ndrivers}_${napis}/${project}/iter_${i}"
+                    DRIVER_FOLDER="${PROJECT_WORKDIR}/drivers"
+                    if [ -d "$DRIVER_FOLDER" ]; then
+                        DRIVER_NAMES="$(find ${DRIVER_FOLDER} -type f -executable)"
+                    else 
+                        continue
+                    fi
+                    # CRASHES=${PROJECT_WORKDIR}/iter_${i}/crashes
+                    
+                    for driver_name in $DRIVER_NAMES; do
+                        driver_name=$(basename $driver_name)
+                        # rm -Rf ${CRASHES}/${driver_name} || true
+                        # mkdir -p ${CRASHES}/${driver_name}
+
+                        DRIVER_CRASHES="${PROJECT_WORKDIR}/crashes/${driver_name}"
+                        if [ "$(ls -A ${DRIVER_CRASHES})" ]; then
+                            # && cp ${DRIVER_CRASHES}/* ${CRASHES}/${driver_name} || 
+
+                            PROJECT_FOLDER="/workspaces/libfuzz/fuzzing_campaigns/workdir_${ndrivers}_${napis}/${project}/iter_${i}"
+                            docker run --privileged --env TARGET=${project} \
+                                    --env TARGET_WORKDIR=${PROJECT_FOLDER} \
+                                    --env GRAMMAR_MODE=${GRAMMAR_MODE} \
+                                    -v $(pwd)/..:/workspaces/libfuzz "${IMG_NAME}-${project}" || true
+                        else
+                            echo "No crashes for ${project}/${driver_name} on iter ${i}"
+                        fi
+                    done
+
+                done
+
+            fi
         done
     done
 done
