+++ fuzzing_campaigns/select_stable_drivers.py
@@ -23,12 +23,13 @@ def _main():
                         help='Driver Folder', required=False)
     parser.add_argument('-threshold', '-t', type=float, default=0.10,
                         help='Threshold for selection', required=False)
-    parser.add_argument('-simulate', '-s', action='store_true',
+    parser.add_argument('-simulate', '-s', choices=['full', 'short'], const='full', nargs='?',
                         help='Simulation only, not moving files around', 
                         required=False)
     parser.add_argument('-timebudget', '-b', type=str,
                         help='Compute budget time for long testing', 
                         required=True)
+    parser.add_argument('-keepcorpus', '-k', action='store_true', help='Keep corpus from previous campaign')
 
     args = parser.parse_args()
 
@@ -37,7 +38,8 @@ def _main():
     simulate = args.simulate
     threshold = args.threshold
     timebudget = normalize_time_boudget(args.timebudget)
-
+    keepcorpus = args.keepcorpus
+    
     libraries = scr.load_report(report, rootdir)
 
     best_drivers = {}
@@ -53,21 +55,22 @@ def _main():
         # print(lib)
         # print(best_drvs)
 
-    if simulate:
+    if simulate in ["full", "short"]:
         print("[INFO] Only simulation, here the drivers I would select:")
         for lib, drvs in best_drivers.items():
             tb = timebudget_per_libary[lib]
             print(f"{lib}: {len(drvs)} drivers w/ timebudget {tb}")
-            for d in drvs:
-                n_drivers = d['n_drivers']
-                n_apis = d['n_apis']
-                driver = d['driver']
-                cov = d['cov']
+            if simulate == "full":
+                for d in drvs:
+                    n_drivers = d['n_drivers']
+                    n_apis = d['n_apis']
+                    driver = d['driver']
+                    cov = d['cov']
 
-                d_path = f"workdir_{n_drivers}_{n_apis}/{lib}/drivers/{driver}"
-                print(f"{d_path}: {cov}")
+                    d_path = f"workdir_{n_drivers}_{n_apis}/{lib}/drivers/{driver}"
+                    print(f"{d_path}: {cov}")
 
-            print("-" * 30)
+                print("-" * 30)
 
         exit(0)
         
@@ -94,9 +97,17 @@ def _main():
             os.system(f"mkdir -p workdir_{n_drivers}_{n_apis}/{lib}/profiles")
             os.system(f"cp {rootdir}/workdir_{n_drivers}_{n_apis}/{lib}/profiles/{driver}_profile workdir_{n_drivers}_{n_apis}/{lib}/profiles")
 
+            # cp cluster driver
+            os.system(f"mkdir -p workdir_{n_drivers}_{n_apis}/{lib}/cluster_drivers")
+            os.system(f"cp {rootdir}/workdir_{n_drivers}_{n_apis}/{lib}/cluster_drivers/{driver}_cluster workdir_{n_drivers}_{n_apis}/{lib}/cluster_drivers")
+
             # cp corpus for driver
             os.system(f"mkdir -p workdir_{n_drivers}_{n_apis}/{lib}/corpus")
-            os.system(f"cp -r {rootdir}/workdir_{n_drivers}_{n_apis}/{lib}/corpus/{driver} workdir_{n_drivers}_{n_apis}/{lib}/corpus")
+            
+            if keepcorpus:
+                os.system(f"cp -r {rootdir}/workdir_{n_drivers}_{n_apis}/{lib}/results/iter_1/corpus_new/{driver} workdir_{n_drivers}_{n_apis}/{lib}/corpus")
+            else:
+                os.system(f"cp -r {rootdir}/workdir_{n_drivers}_{n_apis}/{lib}/corpus/{driver} workdir_{n_drivers}_{n_apis}/{lib}/corpus")
 
             # cp metadata for driver
             os.system(f"mkdir -p workdir_{n_drivers}_{n_apis}/{lib}/metadata")
